Gen 174: Added Integration Test for Product Flow in ADMIN Role and added Unit tests for Product feature (#73)
* Added Unit Tests For Product Service

* InitialCommit

* Added Integartion test to ProductFlow for ADMIN Role

* Added ProductControllerTest test cases for Create,Update and Delete Products

* Resolved Error

+++ src/main/java/com/bank/webApplication/Services/ProductService.java
@@ -13,6 +13,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -28,11 +29,11 @@ public class ProductService {
     public ProductDto getProduct(String productId) {
         log.info("[ProductService] getProduct entered SUCCESS");
         //check if product is present in the database
-        ProductEntity product = productRepository.findById(productId)
-                .orElseThrow(() -> {
-                    log.error("[ProductService] getProduct: not found FAILURE");
-                    return new ProductNotFoundException(" Product Not Found, Invalid Id");
-                });
+        ProductEntity product = productRepository.findByProductId(productId);
+        if (product == null) {
+            log.error("[ProductService] getProduct: not found FAILURE");
+            throw new ProductNotFoundException(" Product Not Found, Invalid Id");
+        }
         //convert to dto from entity
         ProductDto productDto = dtoEntityMapper.convertToDto(product, ProductDto.class);
         log.info("[ProductService] getProduct  SUCCESS");
@@ -59,11 +60,11 @@ public class ProductService {
     public ProductDto createProduct(ProductDto productDto) {
         log.info("[ProductService] createproduct entered SUCCESS");
         //check if the product is already present in the database
-        ProductEntity product = productRepository.findById(productDto.getProductId())
-                .orElseThrow(() -> {
-                    log.error("[ProductService] createProduct: Product Already Exists FAILURE");
-                    return new ProductAlreadyExistException(" Product Already Exists");
-                });
+        ProductEntity product = productRepository.findByProductId(productDto.getProductId());
+        if (product != null) {
+            log.error("[ProductService] createProduct: Product Already Exists FAILURE");
+            throw new ProductAlreadyExistException(" Product Already Exists");
+        }
         //convert from dto to entity
         ProductEntity productEntity = dtoEntityMapper.convertToEntity(productDto, ProductEntity.class);
         log.info("[ProductService] createProduct: New Product created and saved into database");
@@ -75,20 +76,18 @@ public class ProductService {
         //return
         return response;
     }
+
     @Transactional
     //method to updateproduct
     public ProductDto updateProduct(String productId, ProductDto productDto) {
         log.info("[ProductService] updateProduct entered SUCCESS");
         //check if product is present in the database
-        ProductEntity product = productRepository.findById(productId)
-                .orElseThrow(() -> {
-                    log.error("[ProductService] updateProduct: not found FAILURE");
-                    return new ProductNotFoundException(" Product Not Found or does not exist in database");
-                });
-        System.out.println("Before"+product);
-//        System.out.println("After"+product);
+        ProductEntity product = productRepository.findByProductId(productId);
+        if (product == null) {
+            log.error("[ProductService] updateProduct: not found FAILURE");
+            throw new ProductNotFoundException(" Product Not Found or does not exist in database");
+        }
         //populate the updated values into the entity
-//        product.setProductId(productDto.getProductId());
         product.setProductName(productDto.getProductName());
         product.setInterestRate(productDto.getInterestRate());
         product.setFundingWindow(productDto.getFundingWindow());
@@ -98,7 +97,6 @@ public class ProductService {
         log.info("[ProductService] updateProduct:  Product updated and saved into database");
         //save in the database
         ProductEntity update = productRepository.save(product);
-        System.out.println("After update"+update);
         log.info("[ProductService] updateProduct  SUCCESS");
         //convert to dto and return
         return (dtoEntityMapper.convertToDto(update, ProductDto.class));
@@ -108,11 +106,11 @@ public class ProductService {
     public void deleteProduct(String productId) {
         log.info("[ProductService] deleteProduct entered SUCCESS");
         //check if the product is present in the database for deletion
-        ProductEntity product = productRepository.findById(productId)
-                .orElseThrow(() -> {
-                    log.error("[ProductService] Product: not found FAILURE");
-                    return new ProductNotFoundException(" Product Not Found or does not exist in the database");
-                });
+        ProductEntity product = productRepository.findByProductId(productId);
+        if (product == null) {
+            log.error("[ProductService] Product: not found FAILURE");
+            throw new ProductNotFoundException(" Product Not Found or does not exist in the database");
+        }
         log.info("[ProductService] Product Deleted From The Database  SUCCESS");
         //remove product from the database
         productRepository.deleteById(productId);
